+++ content/shell/tools/breakpad_integration_test.py
@@ -14,7 +14,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 import glob
 import json
-import optparse
+import argparse
 import os
 import shutil
 import subprocess
@@ -144,6 +144,37 @@ def run_test(options, crash_dir, symbols_dir, platform,
                                   activity=view_activity, data='chrome://crash',
                                   package=package_name)
     device.StartActivity(launch_intent)
+  elif platform == 'fuchsia':
+    run_test_path = os.path.join(TOP_SRC_DIR, 'build', 'fuchsia', 'test',
+                                 'run_test.py')
+    fuchsia_logs_dir = os.path.join(crash_dir, 'fuchsia_logs')
+    os.makedirs(fuchsia_logs_dir)
+
+    cmd = [
+        sys.executable, run_test_path,
+        '--out-dir', options.build_dir,
+        'content_shell',
+        '--logs-dir', fuchsia_logs_dir,
+        '--',
+        '--run-web-tests',
+        'chrome://crash',
+        '--enable-crash-reporter',
+        '--disable-gpu',
+    ]
+    cmd += additional_arguments
+
+    if options.verbose:
+      print(' '.join(cmd))
+    failure = 'Failed to run content_shell on Fuchsia.'
+    try:
+      # We expect this to fail/crash, so it will likely exit with non-zero.
+      # We capture output to print it if verbose.
+      subprocess.check_output(cmd, stderr=subprocess.STDOUT)
+      print('Warning: run_test.py exited with 0, expected failure/crash.')
+    except subprocess.CalledProcessError as e:
+      if options.verbose:
+        print('run_test.py exited with %d as expected.' % e.returncode)
+        print(e.output.decode('utf-8'))
   else:
     cmd = [options.binary,
            '--run-web-tests',
@@ -166,6 +197,8 @@ def run_test(options, crash_dir, symbols_dir, platform,
   print('# Retrieve crash dump.')
   if platform == 'android':
     minidump = get_android_dump(options, crash_dir, symbols_dir)
+  elif platform == 'fuchsia':
+    minidump = None
   else:
     dmp_dir = crash_dir
     # TODO(crbug.com/41354248): This test should not reach directly into the
@@ -182,7 +215,15 @@ def run_test(options, crash_dir, symbols_dir, platform,
     minidump = dmp_files[0]
 
   print('# Symbolize crash dump.')
-  if platform == 'win32':
+  if platform == 'fuchsia':
+    system_log_path = os.path.join(crash_dir, 'fuchsia_logs', 'system_log')
+    if os.path.exists(system_log_path):
+      with open(system_log_path, 'r') as f:
+        stack = f.read()
+    else:
+      stack = ''
+      print('Warning: system_log not found at %s' % system_log_path)
+  elif platform == 'win32':
     cdb_exe = os.path.join(options.build_dir, 'cdb', 'cdb.exe')
     cmd = [cdb_exe, '-y', options.build_dir, '-c', '.lines;.excr;k30;q',
            '-z', minidump]
@@ -205,7 +246,8 @@ def run_test(options, crash_dir, symbols_dir, platform,
   # Check whether the stack contains a CrashIntentionally symbol.
   found_symbol = 'CrashIntentionally' in stack
 
-  os.remove(minidump)
+  if platform != 'fuchsia':
+    os.remove(minidump)
 
   if options.no_symbols:
     if found_symbol:
@@ -223,29 +265,32 @@ def run_test(options, crash_dir, symbols_dir, platform,
 def main():
   global failure
 
-  parser = optparse.OptionParser()
-  parser.add_option('', '--build-dir', default='',
-                    help='The build output directory.')
-  parser.add_option('', '--binary', default='',
-                    help='The path of the binary to generate symbols for and '
-                         'then run for the test.')
-  parser.add_option('', '--additional-binary', default='',
-                    help='An additional binary for which to generate symbols. '
-                         'On Mac this is used for specifying the '
-                         '"Content Shell Framework" library, which is not '
-                         'linked into --binary.')
-  parser.add_option('', '--no-symbols', default=False, action='store_true',
-                    help='Symbols are not expected to work.')
-  parser.add_option('-j', '--jobs', default=CONCURRENT_TASKS, action='store',
-                    type='int', help='Number of parallel tasks to run.')
-  parser.add_option('-v', '--verbose', action='store_true',
-                    help='Print verbose status output.')
-  parser.add_option('', '--json', default='',
-                    help='Path to JSON output.')
-  parser.add_option('', '--platform', default=sys.platform,
-                    help='Platform to run the test on.')
+  parser = argparse.ArgumentParser()
+  parser.add_argument('--build-dir', default='',
+                      help='The build output directory.')
+  parser.add_argument('--binary', default='',
+                      help='The path of the binary to generate symbols for and '
+                           'then run for the test.')
+  parser.add_argument(
+      '--additional-binary',
+      default='',
+      help='An additional binary for which to generate '
+      'symbols. On Mac this is used for specifying the '
+      '"Content Shell Framework" library, which is not '
+      'linked into --binary.',
+  )
+  parser.add_argument('--no-symbols', default=False, action='store_true',
+                      help='Symbols are not expected to work.')
+  parser.add_argument('-j', '--jobs', default=CONCURRENT_TASKS, type=int,
+                      help='Number of parallel tasks to run.')
+  parser.add_argument('-v', '--verbose', action='store_true',
+                      help='Print verbose status output.')
+  parser.add_argument('--json', default='',
+                      help='Path to JSON output.')
+  parser.add_argument('--platform', default=sys.platform,
+                      help='Platform to run the test on.')
 
-  (options, _) = parser.parse_args()
+  options, unrecognized = parser.parse_known_args()
 
   if not options.build_dir:
     print('Required option --build-dir missing.')
@@ -275,7 +320,7 @@ def main():
       apk_path = os.path.join(options.build_dir, 'apks', 'ContentShell.apk')
       device.Install(apk_path, reinstall=False, allow_downgrade=True)
 
-    if options.platform != 'win32':
+    if options.platform != 'win32' and options.platform != 'fuchsia':
       print('# Generate symbols.')
       bins = [options.binary]
       if options.additional_binary:
@@ -295,7 +340,7 @@ def main():
         failure = 'Failed to run generate_breakpad_symbols.py.'
         subprocess.check_call(cmd)
 
-    run_test(options, crash_dir, symbols_dir, options.platform)
+    run_test(options, crash_dir, symbols_dir, options.platform, unrecognized)
 
   except:
     if failure == '':
@@ -318,6 +363,14 @@ def main():
     if crash_service:
       crash_service.terminate()
       crash_service.wait()
+    if options.platform == 'fuchsia' and 'ISOLATED_OUTDIR' in os.environ:
+      fuchsia_logs_dir = os.path.join(crash_dir, 'fuchsia_logs')
+      if os.path.exists(fuchsia_logs_dir):
+        try:
+          shutil.copytree(fuchsia_logs_dir, os.environ['ISOLATED_OUTDIR'],
+                          dirs_exist_ok=True)
+        except Exception as e:
+          print('Failed to copy fuchsia logs to ISOLATED_OUTDIR: %s' % e)
     try:
       shutil.rmtree(crash_dir)
     except:
